fix(mcp): break ci-health circular import via a leaf types module (#3756)
Importing the MCP tools barrel (or ci-health-check-tool / ci-health-log) under tsx
ESM evaluation threw 'ReferenceError: Cannot access CiHealthStatusSchema before
initialization': the tool DEFINED the shared schema/types and imported the log's
appenders, while the log used the schema at module-eval time — the cyclic init
order left CiHealthStatusSchema in the TDZ. (Forced a static-parse workaround in
the #3687 tool-reference generator.)

Fix: extract CiHealthStatusSchema/CiHealthStatus/CiHealthSignal into a leaf
ci-health-types.ts (imports only zod); both modules import the leaf, breaking the
cycle (check-tool -> log -> types, check-tool -> types — a DAG). The tool re-exports
the symbols for API compat. Verified all three imports load OK under tsx.

Guard: a structural test asserts ci-health-log no longer imports from
ci-health-check-tool + a smoke test loads the tool/barrel. 37 ci-health tests pass;
typecheck, lint, governance, producer-consumer clean.

Closes #3756.
